mqtt_test_client/pyproject.toml
[project]
name = "mqtt-test-client"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "paho-mqtt>=2.1.0",
]

[dependency-groups]
dev = [
    "ruff>=0.15.1",
]

[tool.ruff]
target-version = "py313"
line-length = 120

[tool.ruff.lint]
select = [
    "E", "W", "F", "I", "B", "C4", "UP", "N",
]
ignore = ["E501"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
